add support for multiple boards of same type in OpenOCD boot method
We pass the board id to OpenOCD via a temporary configuration file created
on the fly, so that each board can be identified.

Device types are updated to provide the appropriate board selection
command to OpenOCD, and to disable the gdb port to avoid collisions when
running concurrent jobs on multiple boards.

Signed-off-by: Vincent Wan <vincent.wan@linaro.org>